requirements.txt
openai>=1.50.0
pydantic
Pillow
requests
pdf2image
